matched func_800EA1A8 2

+++ include/gameplay/3CD8.h
@@ -263,10 +263,10 @@ typedef struct _GpCoord64 {
 } GpCoord64;
 STATIC_ASSERT_SIZEOF(GpCoord64, 0x64);
 
-/// 0x10-byte scratch from `G_SCRATCH_HEAD` used by `func_800EA02C`.
-/// `pos` is the low halves of `GsCOORDINATE2.workm.t`. `dir` starts as
-/// `(0, 0x1000, 0)`, is rotated by `D_80070F34`, then added onto `pos`
-/// and passed to `func_800DE7CC`.
+/// 0x10-byte scratch from `G_SCRATCH_HEAD` used by `func_800EA02C` and
+/// `func_800EA1A8`. `pos` is the low halves of the source XYZ. `dir`
+/// starts as `(0, 0x1000, 0)`, is rotated by `D_80070F34`, then added
+/// onto `pos` and passed to `func_800DE7CC`.
 typedef struct _GpRayScratch {
     /* 0x0 */ SVECTOR pos;
     /* 0x8 */ SVECTOR dir;
@@ -308,7 +308,7 @@ void func_800E9BDC(u8 arg0, s32 arg1);
 void func_800E9C6C(void);
 void func_800E9EFC(void);
 s32  func_800EA02C(GsCOORDINATE2* arg0, GsCOORDINATE2* arg1);
-s16  func_800EA1A8(VECTOR3* arg0, VECTOR3* arg1);
+s32  func_800EA1A8(VECTOR3* arg0, VECTOR3* arg1);
 s32  func_800EA318(s16 arg0, s16 arg1, s16 arg2);
 void func_800EA3A0(s32 arg0);
 void func_800EA3B4(void);
